unified message buffer class

+++ server/include/room_manager.h
@@ -8,14 +8,14 @@
 class Room
 {
 public:
-	size_t ID;
+	uint32_t ID;
 	std::string Name;
 
-	size_t ChatGroupId = 0;
+	uint32_t ChatGroupId = 0;
 
 	std::unordered_map<uint32_t, ConnectedClient*> Players;
 
-	Room(size_t id, std::string_view name);
+	Room(uint32_t id, std::string_view name);
 	~Room();
 };
 
@@ -23,7 +23,7 @@ namespace RoomManager
 {
 	void Init();
 	Room* CreateRoom(const std::string& name);
-	Room* GetRoom(size_t id);
+	Room* GetRoom(uint32_t id);
 	Room* GetRoom(const std::string& name);
-	void DestroyRoom(size_t id);
+	void DestroyRoom(uint32_t id);
 }
